# apiv2/internal/docs/swagger.yaml
basePath: /
definitions:
...
  responses.PostResponse:
    properties:
      comments_count:
        type: integer
      created_at:
        type: string
      crop:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_liked:
        type: boolean
      likes_count:
        type: integer
      question:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/responses.UserResponse'
    type: object
  responses.UserResponse:
    properties:
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: ngugijoseph700@gmail.com
    name: API Support
  description: This is the API documentation for the hackatsuki api.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Hackatsuki api
  version: "1.0"
paths:
...
  /posts:
    get:
      description: Retrieve paginated list of posts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'posts: []responses.PostResponse, total_pages: int'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List posts
      tags:
      - Posts
    post:
      consumes:
      - multipart/form-data
      description: Create a new post with question, description, and optional image
      parameters:
      - description: Post question
        in: formData
        name: question
        required: true
        type: string
      - description: Post description
        in: formData
        name: description
        required: true
        type: string
      - description: Crop name
        in: formData
        name: crop
        type: string
      - description: Image file
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a post
      tags:
      - Posts
  /posts/{postId}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Failed to Delete post
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a post
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Retrieve a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single post
      tags:
      - Posts
...
  /posts/{postId}/likes:
    post:
      description: Like or unlike a post by ID (toggles the like status)
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'liked: bool, message: string'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to toggle like on post
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Toggle like on a post
      tags:
      - Posts
...
swagger: "2.0"